fix: SignController() Logic Adjustment
