+++ src/roles/creepBehaviours.ts
@@ -1,4 +1,5 @@
 import { getRoomMemory } from "managers/memoryManager";
+import { PLAYER_USERNAME } from "consts";
 
 /**
  * Shared creep behaviours
@@ -38,4 +39,28 @@ export function addHarvestedEnergy(room: Room, amount: number): void
     const controllerProgressMemory = getRoomMemory(room);
     if (!controllerProgressMemory) return;
     controllerProgressMemory.controllerProgress!.totalEnergyHarvested += amount;
+}
+
+/**
+ * Signs Controller if Required
+ * @param creep 
+ * @param controller 
+ * @param text 
+ * @returns 
+ */
+export function signControllerIfNeeded(creep: Creep, controller: StructureController, text = "007 was here"): boolean
+{
+  // Skip hostile controllers
+  const ownerName = controller.owner?.username;
+  if (ownerName && ownerName !== PLAYER_USERNAME) return false;
+
+  // Only sign if text differs or no sign exists
+  if (controller.sign?.text !== text || controller.sign.username !== PLAYER_USERNAME)
+  {
+    const result = creep.signController(controller, text);
+    if (result === OK) return true;
+    else return false;
+  }
+
+  return false;
 }
